Rename module fburst_mono.py --> fburstpop_mono.py
